Create eric-version1
For the most part, part 1 of the assignment is almost done. I'm getting seg faults whenever i try to free memory thats allocated in the big blocks but allocating memory and freeing smaller blocks is fine. Some of the errors were dealt with but I didn't really get part 2 so that hasn't been done.